+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.7] - 2025-06-26
+
+### Fixed
+- `id`, `attributes` and `relationships` fields optional in resource object, according to jsonapi v1.1#7.2.
+
 ## [0.0.6] - 2025-04-05
 
 ### Fixed
@@ -37,7 +42,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Extract all DTO types from FreeElephants/json-api-php-toolkit to this project
 
-[Unreleased]: https://github.com/FreeElephants/json-api-dto/compare/0.0.6...HEAD
+[Unreleased]: https://github.com/FreeElephants/json-api-dto/compare/0.0.7...HEAD
+[0.0.7]: https://github.com/FreeElephants/json-api-dto/releases/tag/0.0.7
 [0.0.6]: https://github.com/FreeElephants/json-api-dto/releases/tag/0.0.6
 [0.0.5]: https://github.com/FreeElephants/json-api-dto/releases/tag/0.0.5
 [0.0.4]: https://github.com/FreeElephants/json-api-dto/releases/tag/0.0.4